+++ src/process/visitors.rs
@@ -9,8 +9,7 @@ pub trait NodeVisitor<T: NodeProcessor> {
         processor.process_block(block);
 
         block
-            .mutate_statements()
-            .iter_mut()
+            .iter_mut_statements()
             .for_each(|statement| Self::visit_statement(statement, processor));
 
         if let Some(last_statement) = block.mutate_last_statement() {